削除関数用にファイル作成
