+++ features/outputs/functions/delete.ts
@@ -1,0 +1,13 @@
+import axios from "axios";
+import { getApiHeaders } from "@/common/functions/api";
+
+export const callDeleteOutput = async (outputId: number) => {
+  const url = `${process.env.API_ROOT_URL}/api/v1/outputs/${outputId}`;
+  const options = getApiHeaders();
+  try {
+    const response = await axios.delete(url, options);
+    return response.data;
+  } catch (error) {
+    throw new Error(`${JSON.stringify(error)}`);
+  }
+};
